remove feature allowing subs to make import array
This feature makes a subref ambiguous: Is it going to return modules or
imports? The only way to know is to execute the sub, which prevents us
from being able to do interesting things that expect earlier imports to
already be done (like applying roles).

The better way to do this would be to create a subref that also returns
the module name:

    # Old way that no longer works
    our @IMPORT_MODULES = (
        'strict' => sub { [ 'vars' ] },
    );

    # New way that will work
    our @IMPORT_MODULES = (
        sub { return ( strict => [ 'vars' ] ) },
    );