add version check during import with hashref
Sometimes we need to version check the module we're loading. Usually we
can just specify our dependency via CPAN metadata, but sometimes we need
a sanity check, and other times a module can change its behavior based
on the version we asked for.

Now, if you use a hashref as the module name, with the key being the
module, and the value being the version, we will check the module
version.

There is currently a bug in Import::Into that causes this exact
implementation to not work. We could copy the relevant code into our
module, but we'll just wait patiently for the moment.

Refs #25